requirements.txt
streamlit==1.32.0
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.4
plotly==5.19.0
Pillow==10.2.0
pandas==2.2.1
